重构基础结构
改用命名空间的方式进行接口调用
